backend/pyproject.toml
[tool.poetry]
name = "growtree_backend"
version = "0.1.0"
description = "Backend for GrowTree application"
authors = ["TAC"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "3.13.*"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.13.0"
email-validator = "^2.3.0"
bcrypt = "^5.0.0"
jose = "^1.0.0"
fastapi-mail = "^1.6.2"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
